data/ingestion/requirements.txt
azure-storage-blob==12.19.0
azure-data-tables==12.4.4
azure-identity==1.15.0
python-dotenv==1.0.0
requests==2.31.0

